Renaming indra_net -> indranet, etc.
