--- indra/assemblers/indra_net/__init__.py
@@ -1,2 +1,0 @@
-from .indra_net import IndraNet
-from .assembler import IndranetAssembler
