+++ indra/assemblers/indranet/__init__.py
@@ -1,0 +1,2 @@
+from .net import IndraNet
+from .assembler import IndraNetAssembler
